feat: enhance axios response interceptor to handle logout

+++ src/lib/axios.ts
@@ -28,7 +28,8 @@ apiClient.interceptors.response.use(
 
     if (
       error.response?.status === 401 &&
-      !error.config.url?.includes("/auth/login")
+      !error.config.url?.includes("/auth/login") &&
+      !error.config.url?.includes("/auth/logout")
     ) {
       localStorage.removeItem("userData");
       store.dispatch(removeToken());
